se actualiza sincronizar

+++ Scripts/sincronizar.pyw
@@ -37,7 +37,20 @@ def recorte (y):
         prefijo="0"+y[0:4]
     else:
         prefijo=y[0:5]	
- 
+
+def sincronizar (copiar, recibir, diccionario):
+	for i in diccionario:
+				try:
+					shutil.copy(copiar+i, recibir)
+				except:
+					try:			
+						os.remove(recibir+i)
+					except:
+						pass
+
+def diccionario ():
+	pass
+
 #LEER NODOS
 with open (RutaTemp+'nodos.txt','r') as f:
 	nodos=[]
@@ -73,62 +86,51 @@ for nodo in nodos:
 	n=nodo[0] 
 	if direccion != "self":
 #RECIBIR DESDE EL REGISTROS PENDIENTES
-		try:
-			with open (direccion+RutaTemp2+identidad+'.txt','r') as f:
-				g=f.readlines()
-				h=str(g)[2:-3]
-				j=h.split(",")
-				for k in j:
-					recorte(k)
-					prefijo=prefijo.lower()+".pdf"
-					registros3.append(prefijo)
-				documentos3=set(registros3)
-			
-			for i in documentos3:
-				try:
-					shutil.copy(direccion+RutaMadre2+i, RutaMadre)
-				except:
-					try:			
-						os.remove(RutaMadre+i)
-					except:
-						pass	
-			os.remove(direccion+RutaTemp2+identidad+'.txt')    #no escribo registro por que el archivo pendiente solo es para mi, y ya no lo pienso compartir
-		except:
-			pass
+		if os.path.exists(direccion) and os.path.isdir(direccion):
+			try:
+				with open (direccion+RutaTemp2+identidad+'.txt','r') as f:
+					g=f.readlines()
+					h=str(g)[2:-3]
+					j=h.split(",")
+					for k in j:
+						recorte(k)
+						prefijo=prefijo.lower()+".pdf"
+						registros3.append(prefijo)
+					documentos3=set(registros3)
+				
+				sincronizar (direccion+RutaMadre2, RutaMadre, documentos3)			
+				os.remove(direccion+RutaTemp2+identidad+'.txt')    #no escribo registro por que el archivo pendiente solo es para mi, y ya no lo pienso compartir
+			except:
+				pass
 
 #RECIBIR DESDE LOS REGISTROS LOCALES 
 for nodo2 in nodos:
 	direccion=nodo2[1]
 	n2=nodo2[0] 
 	if direccion != "self":
-		try:
-			with open (direccion+RutaTemp2+'registro.txt','r') as ff:
-				g=ff.readlines()
-				h=str(g)[2:-3]
-				j=h.split(",")
-				for k in j:
-					recorte(k)
-					prefijo=prefijo.lower()+".pdf"
-					registros4.append(prefijo)
-				documentos4=set(registros4)
-			for i in documentos4:
-				try:
-					shutil.copy(direccion+RutaMadre2+i, RutaMadre)
-				except:
-					try:			
-						os.remove(RutaMadre+i)
-					except:
-						pass	
-				if i != ".pdf":
-					for nod in vecinos:						#guardo un registro personalizado para mis nodos vecinos, excepto el nodo de origen
-						if nod != n2:
-							with open(RutaTemp+nod+'.txt','a') as q:
-								q.write(i+",")
-				i = ".pdf"
-			ww=open(direccion+RutaTemp2+'registro.txt','w')
-			ww.close()
-		except:
-			pass
+		if os.path.exists(direccion) and os.path.isdir(direccion):
+			try:
+				with open (direccion+RutaTemp2+'registro.txt','r') as ff:
+					g=ff.readlines()
+					h=str(g)[2:-3]
+					j=h.split(",")
+					for k in j:
+						recorte(k)
+						prefijo=prefijo.lower()+".pdf"
+						registros4.append(prefijo)
+					documentos4=set(registros4)
+
+				sincronizar (direccion+RutaMadre2, RutaMadre, documentos4)	
+				for i in documentos4:
+					if i != ".pdf":
+						for nod in vecinos:						#guardo un registro personalizado para mis nodos vecinos, excepto el nodo de origen
+							if nod != n2:
+								with open(RutaTemp+nod+'.txt','a') as q:
+									q.write(i+",")
+				ww=open(direccion+RutaTemp2+'registro.txt','w')
+				ww.close()
+			except:
+				pass
 
 #LEER REGISTRO LOCAL 
 with open (RutaTemp+'registro.txt','r') as l:
@@ -146,50 +148,45 @@ for nodo in nodos:
 	direccion=nodo[1]
 	n=nodo[0]
 	if direccion != "self":	
-		Ruta1=direccion+complemento
-		Ruta2=direccion
-		#EMITIR DESDE EL REGISTRO PENDIENTE
-		try:
-			with open (RutaTemp+n+'.txt','r') as f:
-				g=f.readlines()
-				h=str(g)[2:-3]
-				j=h.split(",")
-				for k in j:
-					recorte(k)
-					prefijo=prefijo.lower()+".pdf"
-					registros2.append(prefijo)
-				documentos2=set(registros2)
-				for i in documentos2:
+		if os.path.exists(direccion) and os.path.isdir(direccion):
+			Ruta1=direccion+complemento
+			Ruta2=direccion
+			#EMITIR DESDE EL REGISTRO PENDIENTE
+			try:
+				with open (RutaTemp+n+'.txt','r') as f:
+					g=f.readlines()
+					h=str(g)[2:-3]
+					j=h.split(",")
+					for k in j:
+						recorte(k)
+						prefijo=prefijo.lower()+".pdf"
+						registros2.append(prefijo)
+					documentos2=set(registros2)
+					sincronizar (RutaMadre, Ruta1, documentos2)
+				os.remove(RutaTemp+n+'.txt')
+			except:
+				pass
+			#EMITIR DESDE REGISTRO LOCAL
+			try:
+				if master=="si":
 					try:
-						shutil.copy(RutaMadre+i, Ruta1)
+						shutil.copy(excel, Ruta2)
 					except:
-						try:			
-							os.remove(Ruta1+i)
-						except:
-							pass	
-			os.remove(RutaTemp+n+'.txt')
-		except:
-			pass
-		#EMITIR DESDE REGISTRO LOCAL
-		try:
-			if master=="si":
-				shutil.copy(excel, Ruta2)		#si soy administrador voy a copiar mi excel en todas partes
-			for i in documentos:
-				shutil.copy(RutaMadre+i, Ruta1)
-				try:
-					shutil.copy(RutaMadre+i, Ruta1)
-				except:
-					try:			
-						os.remove(Ruta1+i)
-					except:
-						pass	
-		except:
+						pass		#si soy administrador voy a copiar mi excel en todas partes
+				sincronizar (RutaMadre, Ruta1, documentos)
+			except:
+				cadena=",".join(documentos)
+				if cadena != ".pdf":
+					pymsgbox.alert("no se pudo sincronizar en "+n,title='Sincronizar Documentos')
+					with open (RutaTemp+n+'.txt','a') as f:
+						f.write(cadena+",")
+		else:	
 			cadena=",".join(documentos)
 			if cadena != ".pdf":
 				pymsgbox.alert("no se pudo sincronizar en "+n,title='Sincronizar Documentos')
 				with open (RutaTemp+n+'.txt','a') as f:
 					f.write(cadena+",")
-				
+
 m = open (RutaTemp+'registro.txt','w')
 m.close()
 pymsgbox.alert("Procedimiento completado",title='Sincronizar Documentos')
